+++ src/main/java/br/com/eric/telegram/kerobot/action/reminder/delete/DeleteReminderAction.java
@@ -20,8 +20,8 @@ public class DeleteReminderAction extends Action {
 
 	@Override
 	public void execute(MessageModel update, int patternPosition, Matcher matcher) {
-		Integer chatId = update.getChat().getId();
-		Integer userId = update.getFrom().getId();
+		Long chatId = update.getChat().getId();
+		Long userId = update.getFrom().getId();
 		deleteReminderExecutor.deleteLast(chatId, userId);
 	}
