+++ src/main/java/br/com/eric/telegram/kerobot/daos/HourInfoRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.com.eric.telegram.kerobot.models.HourInfo;
 
-public interface HourInfoRepository extends CrudRepository<HourInfo, Integer> {
+public interface HourInfoRepository extends CrudRepository<HourInfo, Long> {
 
-	HourInfo findOneByUserId(Integer userId);
+	HourInfo findOneByUserId(Long userId);
 
 }
