+++ src/main/java/br/com/eric/telegram/kerobot/telegram/TelegramApiExecutor.java
@@ -13,15 +13,15 @@ public class TelegramApiExecutor {
 	@Autowired
 	private TelegramApi api;
 
-	public MessageResponse sendMessage(Integer chat_id, String text) {
+	public MessageResponse sendMessage(Long chat_id, String text) {
 		return api.sendMessage(chat_id, text);
 	}
 	
-	public MessageResponse sendMessage(Integer chat_id, String text, InlineKeyboardMarkup inlineKeyboardMarkup) {
+	public MessageResponse sendMessage(Long chat_id, String text, InlineKeyboardMarkup inlineKeyboardMarkup) {
 		return api.sendMessage(chat_id, text, inlineKeyboardMarkup);
 	}
 
-	public MessageResponse sendMessageOrEditMessage(Integer chat_id, String text, InlineKeyboardMarkup inlineKeyboardMarkup, MessageType type, Integer messageId, String callBackQueryId) {
+	public MessageResponse sendMessageOrEditMessage(Long chat_id, String text, InlineKeyboardMarkup inlineKeyboardMarkup, MessageType type, Long messageId, String callBackQueryId) {
 		if (type == MessageType.MESSAGE) {
 			return api.sendMessage(chat_id, text, inlineKeyboardMarkup);
 		} 
@@ -34,15 +34,15 @@ public class TelegramApiExecutor {
 		return response;
 	}
 	
-	public MessageResponse sendVideo(Integer chat_id, String videoURL) {
+	public MessageResponse sendVideo(Long chat_id, String videoURL) {
 		return api.sendVideo(chat_id, videoURL);
 	}
 
-	public MessageResponse sendPhoto(Integer chat_id, String photoURL) {
+	public MessageResponse sendPhoto(Long chat_id, String photoURL) {
 		return api.sendPhoto(chat_id, photoURL);
 	}
 
-	public MessageResponse editMessage(Integer chat_id, Integer messageId, String text) {
+	public MessageResponse editMessage(Long chat_id, Long messageId, String text) {
 		return api.editMessageText(chat_id, text, messageId);
 	}
 }
